more vc++ compiler warning cleanup

+++ source/glest_game/type_instances/unit.cpp
@@ -5237,7 +5237,7 @@ Checksum Unit::getCRC() {
 
     //Commands commands;
 	if(commands.empty() == false) {
-		crcForUnit.addInt(commands.size());
+		crcForUnit.addInt((int)commands.size());
 		for(Commands::const_iterator it= commands.begin(); it != commands.end(); ++it) {
 			uint32 crc = (*it)->getCRC().getSum();
 			crcForUnit.addBytes(&crc,sizeof(uint32));
